error_center/pyproject.toml
[tool.poetry]
name = "error_center"
version = "0.1.0"
description = ""
authors = ["renantamashiro"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
